requirements.txt
wxpython
jpype1
hpyculator